backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
email-validator==2.2.0
alembic==1.13.2
pydantic-settings==2.4.0
python-multipart==0.0.9
